Added methods to manipulate Scene data & moved GlobalTime
